+++ test/dummy/db/migrate/20260804120000_add_resolved_model_to_layered_assistant_messages.layered_assistant.rb
@@ -1,0 +1,6 @@
+# This migration comes from layered_assistant (originally 20260804000000)
+class AddResolvedModelToLayeredAssistantMessages < ActiveRecord::Migration[8.0]
+  def change
+    add_column :layered_assistant_messages, :resolved_model, :string
+  end
+end